As my first Java project I wrote a program that reads two databases, pulls data from them and automatically sends data to their receivers.
Real life project for a non-for-profit organization. Weekly they have a
school to improve public speaking. Every week the students receive
practical assignments where they work on a specific quality.

The program reads the assignments from one of the databases and sends
an email to the person with all the information about the assignment.

For detailed info open the documentation.